requirements.txt
flask>=3.0.0
Pillow>=10.0.0
numpy>=1.26.0
Werkzeug>=3.0.0
pytest>=7.0.0
pytest-cov>=4.0.0